Fix type and unset variables (#13)
* Make the extension from system to user

The options (e.g. width, height) are configured at the user level.
Making the extension enabled at the system level prevents normal users
to change these options.

* Consider unset youtube variables in configuration

User might not configured the youtube extension. Consequently, the
Youtube $user_conf options might not be set yet, raising some errors:

    [warning] --- yt_nocookie does not exist in configuration
    [warning] --- yt_player_width does not exist in configuration
    ...

PHP empty function checks if the variable is declared and if it's not
empty value.

* Register translations when extension is disabled

When the extension is not enabled, the `init` method is not called. In
this case, locales weren't loaded and the configure panel wasn't
translated.

* Release v0.9